+++ tests_r3/test_group00.py
@@ -2,20 +2,15 @@ import unittest
 import cantools
 from cantools.database.can.signal import Signal
 from cantools.database.conversion import BaseConversion
-import run_tests
+import util
+
+db = cantools.database.load_file("r3.dbc")
+message = db.get_message_by_name("group00")
 
 class TestGroup00(unittest.TestCase):
-    db = None
-
-    def setUp(self):
-        if(self.db == None):
-            self.db = cantools.database.load_file("r3.dbc")
-
-        self.message = self.db.get_message_by_name("group00")
-
     def test_rpm(self):
-        run_tests.compare_signals(
-            self.message.signals[0],
+        util.compare_signals(
+            message.signals[0],
             Signal(
                 name="rpm",
                 start=-1+8,
@@ -33,8 +28,8 @@ class TestGroup00(unittest.TestCase):
         )
 
     def test_manifold_pressure(self):
-        run_tests.compare_signals(
-            self.message.signals[1],
+        util.compare_signals(
+            message.signals[1],
             Signal(
                 name="manifold_pressure",
                 start=-1+24,
@@ -52,8 +47,8 @@ class TestGroup00(unittest.TestCase):
         )
 
     def test_throttle_position(self):
-        run_tests.compare_signals(
-            self.message.signals[2],
+        util.compare_signals(
+            message.signals[2],
             Signal(
                 name="throttle_position",
                 start=-1+40,
@@ -71,8 +66,8 @@ class TestGroup00(unittest.TestCase):
         )
 
     def test_coolant_pressure(self):
-        run_tests.compare_signals(
-            self.message.signals[3],
+        util.compare_signals(
+            message.signals[3],
             Signal(
                 name="coolant_pressure",
                 start=-1+56,
